Improved notation consistency on invgauss pdf
I am trying to better understand the parameterization used in this function, so every bit of improved consistency helps. The PDF equation here uses both \mu and m, and later basically clarifies that \mu = m, so I think it's best to use only one of them in the equation. I'm still not convinced that \sigma = s, but that's topic for a different issue/PR.